rename files as client and service instead of complete name
